% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anomaly_lyocv.R
\name{anomaly_lyocv}
\alias{anomaly_lyocv}
\title{Leave Year Out CV model}
\usage{
anomaly_lyocv(
  choose_days_ahead = 21,
  mslp_lag = 30,
  plotting = FALSE,
  dt_all = dt_all,
  form = "anomaly_future ~ FL_anomaly + FL_anomaly_30 + as.factor(month) * nao1_past"
)
}
\arguments{
\item{choose_days_ahead}{Integer. Number of days ahead we are predicting.}

\item{mslp_lag}{Integer. How many days the weather data are lagged}

\item{plotting}{Boolean. If TRUE it plots Skill plot and Pred vs actual for Anomaly and ERA-FL1}

\item{dt_all}{data.table with input data.}

\item{form}{String. Regression formula.}
}
\value{
data.table
}
\description{
Leave Year Out CV model
}
\examples{
mod_21_30_inter_nao = anomaly_lyocv(21, 30,form = "anomaly_future ~ FL_anomaly + FL_anomaly_30 + as.factor(month) * nao1_past")
}
