% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demand_forecast.R
\name{demand_forecast}
\alias{demand_forecast}
\title{Rolling CV forecast model for energy demand.}
\usage{
demand_forecast(
  X_mat1,
  date_demand1,
  forc_start1,
  forc_end1,
  pred_win1 = 30,
  pred_lag1 = 15,
  train_y1 = 5,
  reg_form1,
  p_comps1,
  other_mods1 = NULL,
  comb1 = TRUE,
  custom1 = FALSE,
  incl_climatology1 = TRUE,
  cores = 4
)
}
\arguments{
\item{forc_start}{Date for first initalization date.}

\item{forc_end}{Date for last initialization date.}

\item{pred_win}{Span of days in prediction window, default is 30.}

\item{pred_lag}{Days after initialization day the prediction window begins, default is 15.}

\item{train_y}{Years of training data.}

\item{reg_form}{Regression formula before PC covariates have been added.}

\item{p_comps}{Integer. Number of pc-comps in model.}

\item{other_mods}{List of other models to run or default is NULL.}

\item{comb}{Boolean. FALSE allows specific combination of PCs to be run without running all combinations.}

\item{custom}{Boolean.Specifies a custom pc bases model.}

\item{incl_climatology}{Boolean.Includes climatology model.}
}
\value{
A data.table with actual and predicted demand values.
}
\description{
Rolling CV forecast model for energy demand.
}
\examples{
 Mod1 = demand_forecast(forc_start ='2014-01-01',forc_end = '2021-03-01',pred_win = 30, train_y = 5, pred_lag = 15,reg_form = "volume ~ as.factor(hour) + as.factor(month) + year", p_comps = 3, other_mods= NULL,comb = TRUE)


}
