% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demand_forecast_final.R
\name{demand_forecast_final}
\alias{demand_forecast_final}
\title{Rolling CV forecast model for energy demand.}
\usage{
demand_forecast_final(
  X_mat,
  date_demand,
  NWP_pred,
  forc_start,
  forc_end,
  pred_win = 31,
  pred_lag = 0,
  train_y = 3,
  reg_form,
  p_comps,
  other_mods = NULL,
  comb = TRUE,
  custom = FALSE,
  incl_climatology = FALSE,
  no_pc = TRUE,
  cores = 20,
  setup = "Rolling_final_ensamble"
)
}
\arguments{
\item{forc_start}{String. Date for first initalization date.}

\item{forc_end}{String. Date for last initialization date.}

\item{pred_win}{Integer. Span of days in prediction window, default is 30.}

\item{pred_lag}{Integer. Days after initialization day the prediction window begins, default is 15.}

\item{train_y}{Integer. Years of training data.}

\item{reg_form}{String. Regression formula before PC covariates have been added.}

\item{p_comps}{Integer. Number of pc-comps in model.}

\item{other_mods}{List of other models to run or default is NULL.}

\item{comb}{Boolean. FALSE allows specific combination of PCs to be run without running all combinations.}

\item{custom}{Boolean. Specifies a custom pc based model.}

\item{incl_climatology}{Boolean. Includes climatology model.}
}
\value{
A data.table with actual and predicted demand values.
}
\description{
Rolling CV forecast model for energy demand.
}
\examples{
 Mod1 = demand_forecast_final(forc_start ='2016-01-01',forc_end = '2021-03-01',pred_win = 30, train_y = 3, pred_lag = 0,
reg_form = "volume ~ as.factor(hour) + as.factor(month) + year", p_comps = 3, other_mods= NULL,comb = TRUE)
}
