% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_all_NWP_quantiles.R
\name{get_all_NWP_quantiles}
\alias{get_all_NWP_quantiles}
\title{Get all NWP quantiles
This is a cover function where one inputs the range of forecasts months.
The function then iterates over the specified range calling get_one_month_NWP_quantiles.R,
and stores the results of these iterated outputs in a data.table.
Output can then be merged with ERA-data using a get_ERA_NWP function to be used in prediction tasks.}
\usage{
get_all_NWP_quantiles(
  path = "/mn/kadingir/datascience_000000/eirikhsj/sfe_nordic_temperature/",
  pattern = "sfe_nordic_temperature_",
  start_month = "01",
  start_year = 1993,
  forc_months = 362,
  PC_ERA_path = "/mn/kadingir/datascience_000000/eirikhsj/PC_ERA_79_92.Rda",
  pc_comp = 1,
  reweight = FALSE,
  rew_int = c(15, 4)
)
}
\arguments{
\item{path}{Path to NWP data}

\item{start_month}{Start month.}

\item{start_year}{Start year.}

\item{forc_months}{Number of forecast months from start (including start month)}

\item{pc_comp}{Integer. Number of pca components.}

\item{reweight}{Boolean.}

\item{PC_ERA}{List of matrices containing PC of ERA temperature.}
}
\value{
Returns quantiles of interest
}
\description{
Get all NWP quantiles
This is a cover function where one inputs the range of forecasts months.
The function then iterates over the specified range calling get_one_month_NWP_quantiles.R,
and stores the results of these iterated outputs in a data.table.
Output can then be merged with ERA-data using a get_ERA_NWP function to be used in prediction tasks.
}
\examples{
NWP_quantiles_93_93 = get_all_NWP_quantiles(path = "~/Desktop/MasterNR/Prob2_data",start_month  = '01', start_year = 1993, forc_months = 4,
PC_ERA = PC_ERA_79_92, pc_comp = 2,reweight = FALSE)
}
