% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_one_month_NWP_quantiles.R
\name{get_beta_weights}
\alias{get_beta_weights}
\title{get beta weights
Function which computes the reweighting of the NWP quantiles.}
\usage{
get_beta_weights(k, pc_nwp, ERA_PC1_rew, init_day, target_date, rew_int, sq)
}
\arguments{
\item{k}{Integer. number in sequence of tuning values.}

\item{pc_nwp}{NWP-data for 1 PC to be weighted.}

\item{ERA_PC1_rew}{ERA data.}

\item{init_day}{String. Init day.}

\item{target_date}{Used to store results.}

\item{rew_int}{Day and length of reweighting.}

\item{sq}{sequence of k's.}
}
\value{
data.table
}
\description{
get beta weights
Function which computes the reweighting of the NWP quantiles.
}
\examples{
get_weight(pc_data= pc_data, ERA_PC1_rew =ERA_PC1_rew,init_day=init_day,target_date= target_date, sq = sq, rew_int = rew_int)
}
