% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ERA_NWP.R
\name{get_ERA_NWP}
\alias{get_ERA_NWP}
\title{Loading ERA and NWP data}
\usage{
get_ERA_NWP(
  ERA_path = "~/Desktop/Master2023/Data/PC_ERA/PC_ERA_79_92.Rda",
  NWP_path = "~/Desktop/Master2023/Data/NWP_quant/NWP_quant_1993_2022.Rda",
  quant = "90",
  PC = "PC1",
  NWP_preds = 2,
  reweight = FALSE,
  tuning_k = 10
)
}
\arguments{
\item{ERA_path}{String - Path to ERA file}

\item{NWP_path}{String - Path to NWP file}

\item{PC}{String - Principle component}

\item{NWP_preds}{Integer - Number of NWP predictors}

\item{NWP_quant}{String - Name NWP quantile of interest. Uses the form: NWP_PC1_q90}
}
\value{
Data.table with data ready for input in rolling_mod function
}
\description{
This function matches ERA and NWP data and readies data for use in prediction task in rolling_mod function.
ERA contains the PC of ERA observed temperature (e.g. PC1) .
NWP contains the PC of NWP forecasted temperature (e.g. NWP_1) .
In rolling_mod we are building a rolling cross validation of PC1 ~ f(time) + NWP_PC1_t1 + NWP_PC1_t2
}
\examples{
ERA_NWP = get_ERA_NWP(NWP_preds = 2,reweight = FALSE)


}
