% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_one_month_NWP_quantiles.R
\name{get_one_month_NWP_quantiles}
\alias{get_one_month_NWP_quantiles}
\title{Get quantiles function}
\usage{
get_one_month_NWP_quantiles(
  files_i =
    "/mn/kadingir/datascience_000000/eirikhsj/sfe_nordic_temperature/sfe_nordic_temperature_2008_11.nc4",
  PC_ERA,
  pc_comp,
  rew = FALSE
)
}
\arguments{
\item{PC_ERA}{List. The PC_ERA used to get the NWP factor loadings.}

\item{pc_comp}{Integer. The principle component of interest for}

\item{rew}{Boolean. Whether or not to find reweighted quantiles.}

\item{i}{Integer. File number of forecast month}
}
\value{
Data.table with quantiles 0.1-0.9 and mean of NWP PC of choice.
}
\description{
Finds the correct file of NWP forecast month, gets PCA of forecasts,
and fetches quantiles (0.1-0.9) of 1 selected principle component.
Used in cover-function get_all_NWP_quantiles.R
Specifying reweight = TRUE will also provide reweighted quantiles.
}
\examples{
files = list.files(path = "~/Desktop/Master2023/Data/NWP_monthly/", pattern = '*.nc4', full.names=TRUE)
get_one_month_NWP_quantiles(files_i = 'Desktop/Master2023/Data/NWP_monthly/forecast_2022_07.nc4', PC_ERA = PC_ERA_79_92, pc_comp = 1)

}
