% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_pca.R
\name{get_pca}
\alias{get_pca}
\title{Multi-use function for getting principal-components.}
\usage{
get_pca(X_mat, I_train, I_test, p_comps, NWP = NA, U = NA, mu = NA)
}
\arguments{
\item{X_mat}{Temperature grid matrix.}

\item{I_train}{Indexes of training set.}

\item{I_test}{Indexes of test set.}

\item{p_comps}{Number of principle components to return.}

\item{NWP}{NWP matrix (dim  = 22 21 500 202)}

\item{U}{Use pre-trained U matrix to speed up time. If supplied we do not get dt_train or dt_test only the NWP PC matrix.}

\item{mu}{Pre-trained mean vector, corresponding to U.}
}
\value{
Training (dt_train) and test (dt_test) sets with pca components of ERA temperature field and matching time covariates.
U-matrix (462x462), and mu-vector (1xnrow(X_mat)). NWP-matrix with pca of NWP temperature forecasts.
}
\description{
Fetches the desired number of PCs and returns these within a data.table together with time covariates.
If given an NWP-matrix, the U and mu from the decomposition of ERA-temperature (X_mat) is used to compute and select factor loading of NWP forecasts.
If U and mu are pre-supplied the computation of NWP-forecasts is greatly sped up.
The function is used iteratively in demand_forecast.R
X_mat is supplied by prep_demand_temp_data.R, while I_train and I_test are normally chosen by the rolling cv function (e.g. demand_forecast.R).
}
\examples{
get_pca(X_mat, I_train, I_test, p_comps)


}
