% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_pre_trained_PC_ERA.R
\name{get_pre_trained_PC_ERA}
\alias{get_pre_trained_PC_ERA}
\title{Gets PCA matrix of temperature data for a specified time period.}
\usage{
get_pre_trained_PC_ERA(
  file = "~/Desktop/Master2023/Data/PC_ERA",
  X_mat,
  date_demand = date_demand,
  start_train = "2013-01-01",
  stop_train = "2013-12-31",
  start_test = "2014-01-01",
  stop_test = "2014-12-31",
  NWP = NA,
  run_again = FALSE
)
}
\arguments{
\item{file}{String. File to store and look for PCA matrix.}

\item{X_mat}{ERA temperature matrix.}

\item{date_demand}{Date and demand matrix. We only care about the dates in this application.}

\item{start_train}{String. Start date training.}

\item{stop_train}{String. End date training.}

\item{start_test}{String. Start date test.}

\item{stop_test}{String. End date test.}

\item{NWP}{NWP matrix.}

\item{run_again}{Boolean. If FALSE and file exist function loads pre-made pca matrix.}
}
\value{
A list of data.tables
}
\description{
Gets PCA matrix of temperature data for a specified time period.
}
\examples{
get_pre_trained_PC_ERA(file='~/Desktop/Master2023/Data/PC_ERA', X_mat,
start_train = "1979-01-01", stop_train="1992-12-31", start_test = "1993-01-01", stop_test="2023-01-31",
NWP=NA,run_again =TRUE)


}
