% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_one_month_NWP_quantiles.R
\name{get_weights}
\alias{get_weights}
\title{get weights
Function which computes the reweighting of the NWP quantiles.}
\usage{
get_weights(k, pc_data, ERA_PC1_rew, init_day, target_date, rew_int, sq)
}
\arguments{
\item{k}{Integer. number in sequence of tuning values.}

\item{pc_data}{NWP-data to be weighted.}

\item{ERA_PC1_rew}{ERA data.}

\item{init_day}{String. Init day}

\item{target_date}{Used to store results}

\item{rew_int}{Day and length of reweighting}

\item{sq}{sequence of ks}
}
\description{
get weights
Function which computes the reweighting of the NWP quantiles.
}
