% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pinball_loss.R
\name{pinball_loss}
\alias{pinball_loss}
\title{Pinball loss}
\usage{
pinball_loss(quantile, pred, obs)
}
\arguments{
\item{quantile}{Quantile of interest}

\item{pred}{Predicted values}

\item{obs}{Observed values}
}
\value{
Pinball Loss
}
\description{
Pinball loss
}
\examples{
pinball_loss(quantile = 0.9, pred = 2, obs = 2)

}
