% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_demand_temp_data.R
\name{prep_demand_temp_data}
\alias{prep_demand_temp_data}
\title{Finds, checks and loads temperature and demand data.}
\usage{
prep_demand_temp_data(include_na_volume = TRUE)
}
\arguments{
\item{include_na_volume}{Boolean, include dates with temperature data but na volume data.}
}
\value{
Returns a matrix with grid temperature data and a data.table with demand data.
}
\description{
Finds, checks and loads temperature and demand data.
}
\examples{
prep = prep_demand_temp_data(include_na_volume = TRUE)
}
