% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demand_forecast.R
\name{Rolling_anomaly}
\alias{Rolling_anomaly}
\title{Basic Model function for Demand prediction
Function which computes demand forecast based on anomalies}
\usage{
Rolling_anomaly(
  i,
  X_mat,
  date_demand,
  init_days,
  pred_win,
  pred_lag,
  train_y,
  reg_form,
  p_comps,
  other_mods,
  comb,
  custom,
  incl_climatology,
  no_pc,
  gam_lasso,
  int
)
}
\arguments{
\item{i}{Integer. Iteration}

\item{pred_win}{Integer. Span of days in prediction window, default is 30.}

\item{pred_lag}{Integer. Days after initialization day the prediction window begins, default is 15.}

\item{train_y}{Integer. Years of training data.}

\item{reg_form}{String. Regression formula before PC covariates have been added.}

\item{p_comps}{Integer. Number of pc-comps in model.}

\item{other_mods}{List of other models to run or default is NULL.}

\item{comb}{Boolean. FALSE allows specific combination of PCs to be run without running all combinations.}

\item{custom}{String. Specifies a custom pc based model.}

\item{incl_climatology}{Boolean. TRUE includes climatology model.}

\item{no_pc}{Boolean. TRUE includes basic (no pc) model.}

\item{gam_lasso}{Boolean. TRUE sets lasso penalty.}

\item{int}{Integer. Selects number of climatology models to test.}
}
\value{
data.table
}
\description{
Basic Model function for Demand prediction
Function which computes demand forecast based on anomalies
}
