% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demand_forecast.R
\name{Rolling_climatology}
\alias{Rolling_climatology}
\title{Test function for Climatology2
Function which computes demand forecast.}
\usage{
Rolling_climatology(
  results,
  dt_train,
  dt_test,
  reg_form,
  intervals = 7,
  basic = TRUE,
  daily_cycle = FALSE,
  weekday_cycle = FALSE
)
}
\arguments{
\item{results}{Integer. Iteration}

\item{dt_train}{Integer. Span of days in prediction window, default is 30.}

\item{dt_test}{Integer. Days after initialization day the prediction window begins, default is 15.}

\item{reg_form}{String. Regression formula before PC covariates have been added.}

\item{intervals}{Integer. Number of model intervals to test.}

\item{basic}{Boolean}

\item{daily_cycle}{Boolean. FALSE allows specific combination of PCs to be run without running all combinations.}

\item{weekly_cycle}{Boolean.}
}
\value{
data.table
}
\description{
Test function for Climatology2
Function which computes demand forecast.
}
