% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rolling_mod.R
\docType{data}
\name{rolling_mod}
\alias{rolling_mod}
\alias{q}
\title{Rolling model}
\format{
An object of class \code{numeric} of length 1.
}
\usage{
q
}
\arguments{
\item{q}{Quantile of interest.}

\item{ERA_NWP}{Data of ERA observations and quantiles of NWP forecasts.}

\item{predictors}{Number of NWP predictors.}

\item{model}{Specify type of model, e.g. qreg, qgam, etc.}

\item{window}{Max number of forecast days from initialization.}

\item{hour_v}{Boolean. Include hour covariate.}

\item{week_v}{Boolean. Include week covariate.}

\item{month_v}{Boolean. Include month covariate.}

\item{year_v}{Boolean. Include year covariate.}
}
\value{
Model output with predictions, loss and model coefficients.
}
\description{
Rolling model
}
\examples{
mod = rolling_mod('2007-01-01', '2022-05-01', 0.9, ERA_NWP, 1, model = 'reg', window = 60, reweight=FALSE)
}
\keyword{datasets}
