% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demand_forecast.R
\name{demand_forecast}
\alias{demand_forecast}
\title{Rolling CV forecast model for energy demand.}
\usage{
demand_forecast(
  X_mat,
  date_demand,
  forc_start,
  forc_end,
  pred_win = 30,
  pred_lag = 15,
  train_y = 5,
  reg_form,
  p_comps,
  other_mods = NULL,
  comb = TRUE,
  custom = FALSE,
  incl_climatology = FALSE,
  no_pc = TRUE,
  gam_lasso = FALSE,
  cores = 4
)
}
\arguments{
\item{forc_start}{Date for first initalization date.}

\item{forc_end}{Date for last initialization date.}

\item{pred_win}{Integer. Span of days in prediction window, default is 30.}

\item{pred_lag}{Integer. Days after initialization day the prediction window begins, default is 15.}

\item{train_y}{Integer. Years of training data.}

\item{reg_form}{String. Regression formula before PC covariates have been added.}

\item{p_comps}{Integer. Number of pc-comps in model.}

\item{other_mods}{List of other models to run or default is NULL.}

\item{comb}{Boolean. FALSE allows specific combination of PCs to be run without running all combinations.}

\item{custom}{Boolean. Specifies a custom pc based model.}

\item{incl_climatology}{Boolean. TRUE includes climatology model.}

\item{no_pc}{Boolean. TRUE includes basic (no pc) model.}

\item{gam_lasso}{Boolean. TRUE sets lasso penalty.}

\item{cores}{Integer. Sets number of parallelized cores, default is 4.}
}
\value{
A data.table with actual and predicted demand values.
}
\description{
Rolling CV forecast model for energy demand.
}
\examples{
 Mod1 = demand_forecast(forc_start ='2014-01-01',forc_end = '2023-01-01',pred_win = 30, train_y = 5, pred_lag = 15,reg_form = "volume ~ as.factor(hour) + as.factor(month) + year", p_comps = 3, other_mods= NULL,comb = TRUE)


}
