% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_demand_temp_data.R
\name{prep_demand_temp_data}
\alias{prep_demand_temp_data}
\title{Finds, checks and loads temperature and demand data.}
\usage{
prep_demand_temp_data(
  include_na_volume = TRUE,
  path = "/mn/kadingir/datascience_000000/eirikhsj/"
)
}
\arguments{
\item{include_na_volume}{Boolean. TRUE includes dates with na volume data. Use TRUE for creating historical PC data}

\item{path}{String File path.}
}
\value{
Returns a matrix with grid temperature data and a data.table with demand data.
}
\description{
Finds, checks and loads temperature and demand data.
}
\examples{
prep = prep_demand_temp_data(include_na_volume = TRUE,path = "~/NR/ClimateFutures/RenewableEnergy/Data/DemandForecasting/")
prep = prep_demand_temp_data(include_na_volume = TRUE,path = "/Users/Eirik/Desktop/Master2023/Data/")
prep = prep_demand_temp_data(include_na_volume = TRUE,path = "/nr/samba/user/esjavik/DemandData/")

}
