% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rolling_mod.R
\name{rolling_mod}
\alias{rolling_mod}
\title{Rolling model}
\usage{
rolling_mod(
  forc_start = as.Date("2007-01-01"),
  forc_end = as.Date("2022-05-01"),
  q,
  ERA_NWP,
  predictors,
  model = "reg",
  window = 60,
  reweight = FALSE,
  hour_v = FALSE,
  week_v = FALSE,
  month_v = FALSE,
  year_v = FALSE,
  incl_climatology = FALSE
)
}
\arguments{
\item{q}{Float. Quantile of interest.}

\item{ERA_NWP}{Data.table. Data of ERA observations and quantiles of NWP forecasts.}

\item{predictors}{Integer. Number of NWP predictors.}

\item{model}{String. Specify type of model, e.g. qreg, qgam, etc.}

\item{window}{Integer. Max number of forecast days from initialization.}

\item{reweight}{Boolean. Use reweighted data.}

\item{hour_v}{Boolean. Include hour covariate.}

\item{week_v}{Boolean. Include week covariate.}

\item{month_v}{Boolean. Include month covariate.}

\item{year_v}{Boolean. Include year covariate.}

\item{incl_climatology}{Boolean. Include climatology model}
}
\value{
Model output with predictions, loss and model coefficients.
}
\description{
Rolling model
}
\examples{
mod = rolling_mod('2007-01-01', '2022-05-01', 0.9, ERA_NWP, 1, model = 'reg', window = 60, reweight=FALSE)
}
