% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rolling_mod_NWP.R
\name{rolling_mod_NWP}
\alias{rolling_mod_NWP}
\title{Rolling Temperature PC model using NWP forecasts}
\usage{
rolling_mod_NWP(
  forc_start = as.Date("2007-01-01"),
  forc_end = as.Date("2023-01-01"),
  q = 0.9,
  ERA_NWP,
  formula = "PC1 ~ 1",
  predictors = 1,
  model = "qreg",
  window = 60,
  reweight = FALSE,
  incl_climatology = FALSE,
  incl_other = FALSE,
  skill_interval = 0,
  cores = 4,
  df_spline = 8
)
}
\arguments{
\item{forc_start}{Date. Quantile of interest.}

\item{forc_end}{Date. Quantile of interest.}

\item{q}{Float. Quantile of interest.}

\item{ERA_NWP}{Data.table. Data of ERA observations and quantiles of NWP forecasts.}

\item{formula}{String. Regression formula.}

\item{predictors}{Integer. Number of NWP predictors.}

\item{model}{String. Specify type of model, e.g. qreg, qgam, etc.}

\item{window}{Integer. Max number of forecast days from initialization.}

\item{reweight}{Boolean. Use reweighted data.}

\item{incl_climatology}{Boolean. Include climatology model.}

\item{cores}{Integer. Number of cores to be used.}

\item{df_spline}{Integer. Number of basis splines.}

\item{incl_others}{Boolean. Include climatology model.}
}
\value{
Model output with predictions, loss and model coefficients.
}
\description{
Rolling Temperature PC model using NWP forecasts
}
\examples{
mod = rolling_mod_NWP('2007-01-01', '2022-05-01', 0.9, ERA_NWP, 1, model = 'qreg', window = 60, reweight=FALSE)
}
