% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rolling_mod_NWP.R
\name{rolling_mod_NWP}
\alias{rolling_mod_NWP}
\title{Rolling model}
\usage{
rolling_mod_NWP(
  forc_start = as.Date("2007-01-01"),
  forc_end = as.Date("2023-01-01"),
  q = 0.9,
  ERA_NWP,
  predictors = 1,
  model = "reg",
  window = 60,
  reweight = FALSE,
  incl_climatology = FALSE,
  formula = "PC1 ~ 1",
  incl_other = FALSE,
  skill_interval = 0,
  cores = 4
)
}
\arguments{
\item{q}{Float. Quantile of interest.}

\item{ERA_NWP}{Data.table. Data of ERA observations and quantiles of NWP forecasts.}

\item{predictors}{Integer. Number of NWP predictors.}

\item{model}{String. Specify type of model, e.g. qreg, qgam, etc.}

\item{window}{Integer. Max number of forecast days from initialization.}

\item{reweight}{Boolean. Use reweighted data.}

\item{incl_climatology}{Boolean. Include climatology model.}

\item{formula}{String. Regression formula.}

\item{cores}{Integer. Number of cores to be used.}

\item{hour_v}{Boolean. Include hour covariate.}

\item{week_v}{Boolean. Include week covariate.}

\item{month_v}{Boolean. Include month covariate.}

\item{year_v}{Boolean. Include year covariate.}
}
\value{
Model output with predictions, loss and model coefficients.
}
\description{
Rolling model
}
\examples{
mod = rolling_mod_NWP('2007-01-01', '2022-05-01', 0.9, ERA_NWP, 1, model = 'reg', window = 60, reweight=FALSE)
}
