% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_one_month_NWP_quantiles.R
\name{whdquantile}
\alias{whdquantile}
\title{whdquantile}
\usage{
whdquantile(x, probs, weights = NA)
}
\arguments{
\item{x}{Input}

\item{probs}{Probabilities}

\item{weights}{Weights}
}
\description{
whdquantile
}
