% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_one_month_NWP_quantiles.R
\name{wquantile.generic}
\alias{wquantile.generic}
\title{wquantile.generic
Function which computes the reweighteing of the NWP quantiles.}
\usage{
wquantile.generic(x, probs, cdf.gen, weights = NA)
}
\arguments{
\item{probs}{Integer or list of quantiles of interest.}

\item{weights}{Add weights.}
}
\description{
wquantile.generic
Function which computes the reweighteing of the NWP quantiles.
}
